User Authorization

+++ laragigs/app/Http/Controllers/ListingController.php
@@ -2,7 +2,10 @@
 
 namespace App\Http\Controllers;
 
+
+
 use App\Models\Listing;
+use App\Models\User;
 use Illuminate\Contracts\Session\Session;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
@@ -62,6 +65,13 @@ class ListingController extends Controller
 
     //Update Listing Data
     public function update(Request $request, Listing $listing) {
+
+      //Mkae sure logged in user is owner
+      if($listing->user_id != auth()->id()){
+        abort(403, 'Unauthorized Action');
+      }
+
+
         $formFields = $request->validate([
           'title' => 'required',
           'company' => ['required'],
@@ -85,7 +95,18 @@ class ListingController extends Controller
 
       //Delete Listing
       public function destory(Listing $listing){
+
+         //Mkae sure logged in user is owner
+      if($listing->user_id != auth()->id()){
+        abort(403, 'Unauthorized Action');
+      }
+      
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
+      }
+
+      //Manage Listings
+      public function manage(){
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
       }
 }
